+++ src/Library.Core/Models/Reservation.cs
@@ -3,11 +3,11 @@ using System.ComponentModel.DataAnnotations.Schema;
 namespace Library.Core.Models
 {
     public class Reservation: BaseEntity {
-        public int PatronId { get; set; }
+        public long PatronId { get; set; }
         public bool Active { get; set; }
         public int NumberOfDaysId { get; set; }
         public int BasisId { get; set; }        
-        public int ReservedByUserId { get; set; }
+        public long ReservedByUserId { get; set; }
 
         [ForeignKey("PatronId")]
         public User Patron { get; set; }
